Pull request 83: add-url-upd
Squashed commit of the following:

commit ee41ef38e76d93df17d9d89cecc7cdac9cb66457
Author: Ainar Garipov <A.Garipov@AdGuard.COM>
Date:   Wed Sep 6 15:04:53 2023 +0300

    urlutil: mv

commit 3f0631781d0ce7c2fd01c3683ad6a2f8c7181333
Author: Ainar Garipov <A.Garipov@AdGuard.COM>
Date:   Wed Sep 6 13:19:05 2023 +0300

    tools: add todo

commit 3f744f445887539cd9c60fa5f3d8e1bdded4fc1d
Author: Ainar Garipov <A.Garipov@AdGuard.COM>
Date:   Wed Sep 6 13:14:47 2023 +0300

    all: upd ci

commit e56df17279cf80ca4423823dd00139cd97317a5b
Author: Ainar Garipov <A.Garipov@AdGuard.COM>
Date:   Wed Sep 6 13:12:13 2023 +0300

    urlutil: add pkg; all: upd

+++ netutil/addr_test.go
@@ -110,7 +110,7 @@ func TestSplitHostPort(t *testing.T) {
 		in         string
 		wantErrMsg string
 		wantHost   string
-		wantPort   int
+		wantPort   uint16
 	}{{
 		name:       "success_ipv4",
 		in:         "1.2.3.4:12345",
